Whats New Apis

+++ controller/tickerController.js
@@ -5,16 +5,22 @@ const response = require('../response')
 
 async function createTicker(req, res, next) {
     try {
-        let file = req.files.file;
-        let { description, hyperLinkUrl} = req.body
+        let file = null;
+        let s3Response = {};
+
+        if (req.files && req.files.file) {
+            file = req.files.file;
+            let fileName = file.name;
+            s3Response = await fileUpload.uploadFile(fileName, file.data, 'ticker')
+        }
+
+        let { description, hyperLinkUrl } = req.body
         const userId = req.headers.userId
-        console.log("file object", file)
-        let fileName = file.name;
-        let s3Response = await fileUpload.uploadFile(fileName, file.data, 'ticker')
+
         console.log('s3Response==============>', s3Response)
         let tickerData = {
             description: description,
-            hyperLinkUrl:hyperLinkUrl,
+            hyperLinkUrl: hyperLinkUrl,
             documentUrl: s3Response.Location,
             s3FileKey: s3Response.Key,
             userId: userId
@@ -62,29 +68,31 @@ async function deleteTicker(req, res, next) {
 
 async function updateTicker(req, res, next) {
     try {
-        let { tickerId, description, hyperLinkUrl,isNewFile } = req.body
-        
+        let { tickerId, description, hyperLinkUrl, isNewFile } = req.body
+
         const userId = req.headers.userId
-        
+
         let tickerWhereCaluse = {
             id: tickerId
         }
         let tickerResult = await tickerRepo.getTickerByCondition(tickerWhereCaluse)
         let updatedS3FileKey = tickerResult.s3FileKey;
         let updatedImgUrl = tickerResult.imgUrl;
-       
-        
+
+
         if (tickerResult) {
             //Deleting Files from s3 Bucket 
             if (isNewFile == 'Y') {
                 let file = req.files.file
                 let fileName = file.name
-               
+
                 //let s3Filekey = updatedS3FileKey;
-                let deleteFileList = [{
-                    Key: updatedS3FileKey
-                }]
-                await fileUpload.deleteFile(deleteFileList, 'ticker')
+                if (updatedS3FileKey !== null && updatedS3FileKey !== '') {
+                    let deleteFileList = [{
+                        Key: updatedS3FileKey
+                    }]
+                    await fileUpload.deleteFile(deleteFileList, 'ticker')
+                }
                 let s3Response = await fileUpload.uploadFile(fileName, file.data, 'ticker')
                 updatedS3FileKey = s3Response.Key;
                 updatedImgUrl = s3Response.Location;
@@ -92,13 +100,13 @@ async function updateTicker(req, res, next) {
 
             let tickerData = {
                 description: description,
-                hyperLinkUrl:hyperLinkUrl,
+                hyperLinkUrl: hyperLinkUrl,
                 documentUrl: updatedImgUrl,
                 s3FileKey: updatedS3FileKey,
                 userId: userId
             }
 
-            
+
 
             let tickerUpdate = tickerRepo.updateTicker(tickerWhereCaluse, tickerData)
             if (tickerUpdate) {
